refactor: P2 server split 4/N — extract /api/skills route domain + pure file helpers
Pure cut-paste, zero behavior change:

- server/file-helpers.ts (new): the stateless helpers shared by route
  domains and server.ts — safeJoin, slugifySkillName, canonicalTreeRoot,
  WorkspaceTreeNode, WORKSPACE_TREE_MAX_DEPTH, TEXT_PREVIEW_EXTENSIONS,
  TEXT_NOEXT_NAMES, OFFICE_PREVIEW_EXTENSIONS, isDotfileText, officeFormat,
  contentTypeForWorkspaceFile, workspaceFileKind
- server/routes/skills.ts (new): all 12 blocks covering /api/skills*
  (list/upload/reload/toggle/delete/content/tree/file/raw) and
  /api/skill-library* (list/import) become handleSkillsRoutes(...):
  Promise<boolean>
- Skill helpers that close over server module state (listProjectSkills,
  setSkillEnabled, installSkillZip/Markdown, listSkillLibrary,
  importSkillFromLibrary) stay in server.ts for now — the workspace upload
  routes still call installSkillZip/Markdown — and are injected via
  SkillsRouteContext. SkillLibraryItem moves to the routes module and is
  re-imported as a type
- Smoke test gains three skills assertions (list 200, upload validation
  400, missing-skill content 404)

server.ts 4411 -> 3985 lines.
